add: scaffold out CRISP contracts
Split up the contracts into two directories `evm-base` and `risc0`. The idea here is that there are probably common elements between CRISP solidity implementations for various compute providers. I imagine the same is true for the Rust implementation, so perhaps it's worth also treating the current `compute_provider` repo as a base that the implementations for each compute provider can inherit from. The `risc0` directory was generated by `forge init` using the risc0 template, so it also includes templated host and guest code.